requirements.txt
yfinance
pip-system-certs
psycopg2-binary==2.9.9
apache-airflow
pyspark